Add "Translocator" class
"Translocator" class use lang file (located in ./lang folder) to translate string.

The class is initialized in game class using the lang folder path.

English is the default language, to change it, please refer to the documentation.

To use it, call qTranslate(char*) or stdTranslate(char*) to get the string in the current language.