Add GraphicsCaroussel class and refactor graphic part
Add GraphicsCaroussel class which display a string and 2 arrows and emit an event when arrows are clicked

Edit some scene to use carousel and button items. All classes has been refactored exept CharacterSelectionMenu.

Add "Restart" and "Quit" buttons to BoardScreen class. For the moment only the "Quit" button is working and return to the main menu.