Add label item and fix bugs between console and graphic interfaces
Add GraphicsLabeltem which is just a QGraphicsTextItem linked to the Textures class.

Fix bugs between graphic and console version. Mixed mode is now allowed.

Add some methods to GameCore class. The game is stopped when there are less than 2 people who can play.

Reduce size of texts in label and buttons.

Rename the function useAccents() to removeAccents() which know return the opposite.