Remove GraphicsLabelItem, add walls and music
Remove unused GraphicsLabelItem class

Add walls to the board which are randomly generated with the board.

Change the return type of volume() method to int instead of QString. Options menu edited to fit the changes.

Add a method to load QUrl of sound file.

Replace semi-transparent pixmap by full transparent pixmap.

Correct a bug when a player click in a unavailable box after clicking on one of his pawn.

Edit the GraphicsPlayerFrameItem. If player var is set to -1, the text is not displayed.

Add pointing cursor when hovering carousel's arrows

Show winner on the board screen and freeze the board.

Refactor the CharacterSelectionMenu.

Add musics and sounds on the graphic version.

Correct a bug in console only mode.