Default resources added
Default resources has been added to the qrc file. They are used only if no textures packs or no language file have been found.

The default language is "English" and the default texture pack is "The Original".

Fonts can now be added to textures packs. If their is a font file name "font.ttf" at the root of the textures pack, this font will be used. If this file doesn't exist, the font specified in the pack.conf will be used.

Rename lang folder to langs